Add ServerMiddleware to notify when worker exceed a configured time
The ServerMiddleware is called from sidekiq server after it has pulled
form the redis workers queue and ready to perform.
NOTE: Not calling yield or returning false or nil will result in no
further middleware being called and the job will not be pushed to the queue,
see https://github.com/mperham/sidekiq/wiki/Middleware

The recorder class will register the worker status and notify if the job work takes
more than the exceeded_maximum_time configured time or if it fails

The default exceeded_maximum_time is 0, so if not configured it always notify the message.